(1866) deport logic from activities.api
